Rename file for main note opener logic
