+++ src/index.ts
@@ -1,7 +1,10 @@
 import { Notice, Plugin } from "obsidian";
 import { appHasDailyNotesPluginLoaded } from 'obsidian-daily-notes-interface';
-import { openTomorrowsDailyNote } from "./OpenTomorrowsDailyNote";
-import { TomorrowsDailyNoteSettingTab, TomorrowsDailyNoteSettings } from "./settings";
+import { openTomorrowsDailyNote } from "./noteOpener";
+import {
+  TomorrowsDailyNoteSettingTab,
+  TomorrowsDailyNoteSettings
+} from "./settings";
 
 const DEFAULT_SETTINGS: Partial<TomorrowsDailyNoteSettings> = {
   skipWeekends: false
