+++ src/noteOpener.ts
